requirements.txt
streamlit
openai== 0.28
pyyaml
pandas
langchain
langchain_community
backoff
langchain_openai